pyproject.toml
[project]
name = "ormah"
version = "0.1.1"
description = "Local-first, portable, LLM-agnostic memory system for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Rishikesh Chirammel Ajit" },
]
keywords = ["memory", "ai", "agents", "mcp", "llm", "knowledge-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-frontmatter>=1.1.0",
    "fastembed>=0.4.0",
    "sqlite-vec>=0.1.6",
    "apscheduler>=3.10.0",
    "watchdog>=6.0.0",
    "mcp>=1.0.0",
    "httpx>=0.28.0",
    "websockets>=14.0",
    "python-slugify>=8.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.8.0",
]
litellm = ["litellm>=1.0.0"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.8.0",
]

[project.scripts]
ormah = "ormah.cli:main"
ormah-mcp = "ormah.adapters.mcp_adapter:main"
ormah-cli = "ormah.adapters.cli_adapter:main"

[project.urls]
Homepage = "https://ormah.me"
Repository = "https://github.com/r-spade/ormah"
Issues = "https://github.com/r-spade/ormah/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
artifacts = ["src/ormah/ui_dist/**"]

[tool.hatch.build.hooks.custom]

[tool.hatch.build.targets.wheel]
packages = ["src/ormah"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
